Split reporter configuration out of all-in-one
The all-in-one config adds it's own "127.0.0.1" outside of our control
which means using a `dns:///` address for external collectors isn't
resolving properly as a `dns:///` entry should be the only entry.
The easiest way around this is to drop support for external collectors
(reporter config) in the all-in-one job as if a user is deploying
external collectors then they can just run the individual components.