Move downsampling options into collector config
These aren't common to all jobs - the agents are expected to run as a
sidecar on every VM so won't handle enough traffic to need downsampling
(I think that's the theory at least). The collectors are fan-in points
so will handle that kind of traffic hence the downsampling options
belong in there.